requirements.txt
# Приложение и работа с jwt
fastapi==0.115.12
python-jose==3.3.0

# ЛОггер
loguru==0.7.3

# ORM, миграции и работа с базой
tortoise-orm==0.25.0
aerich==0.2.4
asyncpg==0.30.0

# Запуск приложения
uvicorn==0.34.2
gunicorn==23.0.0

# Dfkblfwbz
pydantic==2.10.6
pydantic-settings==2.9.1s

# Utils
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.27.2
bcrypt==4.3.0

# Асинхронная почта
aiosmtplib==4.0.1

# Тестирование
pytest-asyncio==0.25.3

# Линтер
ruff==0.11.11

# Метрики (прометей, ошибки)
prometheus_client==0.21.1

# Трейсер (отслежиивание)
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-instrumentation-fastapi==0.52b1
